# v2/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Clovery API
  version: 0.1.0
  description: Versioned contract shared by Clovery clients and the Go API.
paths:
...
  /v1/auth/password/reset/start:
    post:
      operationId: startPasswordReset
      summary: Start a password reset with an approved recovery method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetStartRequest"
      responses:
        "202":
          description: A generic reset intent was created when eligible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetStartResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/password/reset/complete:
    post:
      operationId: completePasswordReset
      summary: Complete a password reset and revoke prior refresh sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetCompleteRequest"
      responses:
        "204":
          description: Password changed and prior refresh sessions revoked.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /v1/account/devices:
    get:
      operationId: listCurrentAccountDevices
      summary: List devices belonging to the authenticated Clovery account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account-scoped device list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/billing/apple/restore:
    post:
      operationId: restoreAppleTransactions
      summary: Reverify Apple transactions and return root-account entitlements
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleRestoreRequest"
      responses:
        "200":
          description: Current root-account entitlements after restore.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntitlementListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          description: At least one Apple transaction failed cryptographic or claim validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Apple verification is temporarily unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/vault/assets/uploads:
    post:
      operationId: startVaultAssetUpload
      summary: Create an idempotent private asset record and a short-lived upload URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUploadRequest"
      responses:
        "201":
          description: Private upload ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/vault/assets/{assetId}/complete:
    post:
      operationId: completeVaultAssetUpload
      summary: Read the uploaded object and verify its byte size and SHA-256 before completion
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "204":
          description: Asset verified and completed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/vault/assets/{assetId}/download:
    get:
      operationId: downloadVaultAsset
      summary: Create a short-lived private download URL for a completed asset
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Private download ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetDownloadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/vault/migrations:
    post:
      operationId: createVaultMigration
      summary: Create an idempotent V1 or legacy CloudKit migration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: The complete encoded JSON request body is limited to 25165824 bytes.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MigrationCreateRequest"
      responses:
        "201":
          description: Migration created or existing matching migration returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AssetId:
      name: assetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The requested identifier or binding is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed without revealing account existence.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated account does not own the requested resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The account-scoped resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Too many attempts; retry after the response delay.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: The request is well-formed but failed integrity or business validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: The operation is temporarily disabled or its dependency is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PasswordResetStartRequest:
      type: object
      additionalProperties: false
      required: [login_id, recovery_method]
      properties:
        login_id:
          type: string
        recovery_method:
          type: string
          enum: [passkey, recovery_email, recovery_code]
    PasswordResetStartResponse:
      type: object
      additionalProperties: false
      required: [accepted, expires_in]
      properties:
        accepted:
          type: boolean
          const: true
        reset_intent_id:
          type: string
          format: uuid
        challenge:
          type: string
          writeOnly: true
        expires_in:
          type: integer
          minimum: 1
    PasswordResetCompleteRequest:
      type: object
      additionalProperties: false
      required: [reset_intent_id, proof, new_password]
      properties:
        reset_intent_id:
          type: string
          format: uuid
        proof:
          type: string
          writeOnly: true
        new_password:
          type: string
          minLength: 12
          maxLength: 256
          writeOnly: true
...
    Device:
      type: object
      additionalProperties: false
      required: [device_id, display_name, platform, created_at, current]
      properties:
        device_id:
          type: string
          format: uuid
        display_name:
          type: string
        platform:
          type: string
          enum: [ios, android, harmonyos]
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: [string, "null"]
          format: date-time
        current:
          type: boolean
    DeviceListResponse:
      type: object
      additionalProperties: false
      required: [devices]
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
...
    AssetUploadRequest:
      type: object
      additionalProperties: false
      required: [asset_id, content_type, byte_size, sha256]
      properties:
        asset_id:
          type: string
          format: uuid
        content_type:
          type: string
          enum: [image/jpeg, image/png, image/heic, image/heif, image/webp]
        byte_size:
          type: integer
          format: int64
          minimum: 1
          maximum: 52428800
        sha256:
          type: string
          pattern: "^[a-f0-9]{64}$"
    AssetUploadResponse:
      type: object
      additionalProperties: false
      required: [asset_id, status]
      properties:
        asset_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [upload_required, complete]
        upload_url:
          type: string
          format: uri
        required_headers:
          type: object
          additionalProperties:
            type: string
        expires_at:
          type: string
          format: date-time
    AssetDownloadResponse:
      type: object
      additionalProperties: false
      required: [asset_id, download_url, expires_at]
      properties:
        asset_id:
          type: string
          format: uuid
        download_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
    MigrationCreateRequest:
      type: object
      additionalProperties: false
      required:
        - migration_id
        - format_version
        - source
        - entry_count
        - asset_count
        - total_bytes
        - manifest_sha256
        - manifest_base64
      properties:
        migration_id:
          type: string
          format: uuid
        format_version:
          type: integer
          enum: [1]
        source:
          type: string
          enum: [v1_bundle, legacy_cloudkit]
        entry_count:
          type: integer
          minimum: 0
        asset_count:
          type: integer
          minimum: 0
        total_bytes:
          type: integer
          format: int64
          minimum: 0
          description: Canonical active-entry payload bytes, two bytes for each deleted-entry tombstone, plus asset bytes.
        manifest_sha256:
          type: string
          pattern: "^[a-f0-9]{64}$"
          description: SHA-256 of the exact decoded manifest.json bytes.
        manifest_base64:
          type: string
          format: byte
          maxLength: 16777216
          description: Exact bytes of manifest.json encoded as base64. The manifest must bind the raw entries/deleted-ID files and every active entry by source ID, SHA-256, and canonical byte length.
    MigrationResponse:
      type: object
      additionalProperties: false
      required:
        - migration_id
        - format_version
        - source
        - entry_count
        - deleted_count
        - asset_count
        - total_bytes
        - manifest_sha256
        - status
        - created_at
      properties:
        migration_id:
          type: string
          format: uuid
        format_version:
          type: integer
        source:
          type: string
        entry_count:
          type: integer
        deleted_count:
          type: integer
          minimum: 0
        asset_count:
          type: integer
        total_bytes:
          type: integer
          format: int64
        manifest_sha256:
          type: string
        status:
          type: string
          enum: [uploading, verified, failed]
        created_at:
          type: string
          format: date-time
...
    AppleEnvironment:
      type: string
      enum: [production, sandbox]
...
    AppleRestoreRequest:
      type: object
      additionalProperties: false
      required: [transaction_ids, environment]
      properties:
        transaction_ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 128
        environment:
          $ref: "#/components/schemas/AppleEnvironment"
    Entitlement:
      type: object
      additionalProperties: false
      required:
        - product_id
        - state
        - source_storefront
        - source_transaction_id
        - updated_at
      properties:
        product_id:
          type: string
        state:
          type: string
          enum: [active, expired, revoked, cancelled, failed]
        expires_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
        source_storefront:
          type: string
        source_transaction_id:
          type: string
        updated_at:
          type: string
          format: date-time
    EntitlementListResponse:
      type: object
      additionalProperties: false
      required: [entitlements]
      properties:
        entitlements:
          type: array
          items:
            $ref: "#/components/schemas/Entitlement"
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...